feature/PT-377

+++ OsmIntegrator/ApiModels/Tiles/UncommitedTile.cs
@@ -22,6 +22,6 @@ namespace OsmIntegrator.ApiModels.Tiles
     [Required]
     public int GtfsStopsCount { get; set; }
     public string AssignedUserName { get; set; }
-    public ushort UnconnectedGtfsStops { get; set; }
+    public int UnconnectedGtfsStops { get; set; }
   }
 }
